Remove warning about 'context' not used
When creating a test method from a `should` declaration, we were using
the string version of `class_eval`. Before the `class_eval` we set a
`context` variable and we were then referring to that variable from
inside the method, inside the string. But the Ruby parser doesn't know
this so it thinks that `context` is not being used. There's no real
reason we need to use a string here, we can use a block instead.