Autoload macros via Railtie
Early versions of Rails shipped with a feature that allowed you to
add a gem to your dependencies which, upon being loaded, would hook into
your app and extend its behavior in some way. As long as the gem
contained a `rails/init.rb` file, Rails would see this and load this
automatically. When Rails 3 came out, this feature was deprecated in
favor of Railties, and when Rails 4 came out, it removed the autoloading
of `rails/init.rb` altogether.

When this happened, this broke a core feature of Shoulda Context, which
allows a developer to define custom macros (i.e. class-level methods
that encapsulate complex assertions) which will automatically get mixed
into the parent test case class (Minitest or Test::Unit). This commit
adds a Railtie to the gem, through which we can accomplish the
autoloading functionality.

Co-authored-by: Mark H. Wilkinson <mhw@dangerous-techniques.com>