Correct rerun snippet in Rails projects
This partially reverts commit 9b791b4f26e49930f59a49c5b2acff67aabcc15f,
and then fixes the use of `context` so that it doesn't generate a
warning.

In addition to fixing the rerun snippet, in writing tests, a
compatibility with Rails::TestUnitReporter was revealed and this was
also fixed as well.