Remove 'assigned but unused variable' warning
In converting a `should` within a `context` into a Minitest test (i.e.
method), we have to make use of `class_eval`. Inside of this eval we
refer to the Context object that we're inside of, which means we have to
set a variable outside of the eval. Unfortunately Ruby can't see into
the eval so it thinks that variable has gone unused and emits a warning.
The only real way to fix this is to make use of the variable somehow.
Double assigning the variable seems to do the trick.

Hat tip: @brandur, @pyrmont