Drop support for Ruby 2.3
I hesitated in doing this, but it's a really old version and modern
versions of Rails do not support it and have not supported it for some
time. It ends up being more trouble than it's worth to support this (it
requires some hacks to the Appraisals file and whatnot).